rag_service/requirements.txt.txt
fastapi
uvicorn
pydantic
python-dotenv
qdrant-client
requests